refactor(tools): split tool-output-content module
